% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gift_details}
\alias{gift_details}
\title{Example of details of gifts received and their senders}
\format{
\subsection{\code{gift_details}}{

A data frame with 4 rows and 6 columns:
\describe{
\item{addr1}{Address line 1}
\item{addr2}{Address line 2}
\item{city}{Address city}
\item{country}{Address country}
\item{name}{Name of the sender}
\item{gift}{Gift description}

}
}
}
\usage{
gift_details
}
\description{
Example details of some gifts received and their senders.
Note that these are trivial examples only, not real records
of gifts received.
}
\keyword{datasets}
