% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gifts}
\alias{gifts}
\title{Example gifts received}
\format{
\subsection{\code{gifts}}{

A data frame with 4 rows and 6 columns:
\describe{
\item{addr1}{First line of the address}
\item{addr2}{Second line of the address}
\item{city}{City of the address}
\item{country}{Country of the address}
\item{name}{Name of the sender of the gift}
\item{gift}{The item received as a gift}
}
}
}
\usage{
gifts
}
\description{
A fictional table of gifts received. Includes the sender, their address
and the gift item.
}
\keyword{datasets}
