% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_placeholders.R
\name{sub_placeholders}
\alias{sub_placeholders}
\title{Subsitute placeholders}
\usage{
sub_placeholders(template_doc, replacements, output_doc)
}
\arguments{
\item{output_doc}{}
}
\description{
Subsitute placeholders
}
