% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_paragraph_if_na.R
\name{delete_paragraph_if_na}
\alias{delete_paragraph_if_na}
\title{Deletes the entire paragraph for the placeholder if the value is NA}
\usage{
delete_paragraph_if_na(doc, placeholder, value, ph_start = "<<", ph_end = ">>")
}
\arguments{
\item{ph_end}{}
}
\description{
Deletes the entire paragraph for the placeholder if the value is NA
}
