% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docmerge.R
\name{docmerge}
\alias{docmerge}
\title{Generate word documents from template and dataframe of inputs}
\usage{
docmerge(template_doc, merge_inputs, output_path = "./")
}
\arguments{
\item{output_path}{}
}
\description{
Generate word documents from template and dataframe of inputs
}
